Write _.last to give the last n elements of array
Alter _.last so it returns the last element of an array when not given
a value for n.

Use a conditional statement to test if n is larger than array.length
and if so, write _.last to return the entire array.

Use the else statement for when the user enters a valid n. In this
case, _.last will return the elements from index n until the end of the
array.