Write _.each to call iterator on values in a collection
Make _.each able to call iterator(value, key, collection) on every
value inside a collection.

Use Array.isArray() to test if the collection is an array type. If so,
use a normal for loop to call iterator on each element. If not, the
collection is assumed to be an object. In this case, _.each will use
the for-in loop to call iterator on each property value in the object.