Write _.uniq and _.map to duplicate/transform
Write _.uniq such that it produces a copy of an array, except without
any duplicate values. Use _.each and pass it an anonymous function that
only pushes a value into a new array if it doesn't already exist in it.

Write _.map such that it returns a new array that contains a
collection's values after they are transformed by the iterator. The
original array is not altered.