Write _.every to see if all values pass a test
Write _.every such that it takes a collection of values and returns
false if any of the values fails a test.

Make it so _.every works with undefined values, casts results to a
boolean, and handle callbacks that manipulate the input. _.every
should work even when no callback function is provided.