Alter _.some to pass all test cases using _.every
Find a way to use _.every in order to make _.some pass all test cases.
Store the result of _.every in toPass. Change the variable pass to
foundMatch, which is set to true when _.every iterates through the
collection and finds a value that passes the iterator's test. Change
the return statement to return foundMatch || toPass.